chore: move ConsensusParameters and GlobalParameters into their own module.
