+++ crates/amaru-kernel/src/lib.rs
@@ -47,6 +47,7 @@ pub use cardano::{
     bytes::{Bytes, empty_bytes},
     certificate::Certificate,
     certificate_pointer::CertificatePointer,
+    consensus_parameters::ConsensusParameters,
     constitution::Constitution,
     constitutional_committee::ConstitutionalCommitteeStatus,
     cost_model::{self, CostModel},
@@ -66,6 +67,9 @@ pub use cardano::{
     era_summary::EraSummary,
     ex_units::{self, ExUnits, sum_ex_units},
     ex_units_prices::{self, ExUnitPrices},
+    global_parameters::{
+        GlobalParameters, MAINNET_GLOBAL_PARAMETERS, PREPROD_GLOBAL_PARAMETERS, PREVIEW_GLOBAL_PARAMETERS,
+    },
     governance_action::GovernanceAction,
     hash::{self, Hash, Hasher, HeaderHash, NULL_HASH28, NULL_HASH32, ORIGIN_HASH, PoolId, size},
     header::Header,
@@ -103,9 +107,8 @@ pub use cardano::{
     proposal_pointer::ProposalPointer,
     proposal_state::ProposalState,
     protocol_parameters::{
-        ConsensusParameters, GlobalParameters, MAINNET_DEFAULT_PROTOCOL_PARAMETERS, MAINNET_GLOBAL_PARAMETERS,
-        PREPROD_DEFAULT_PROTOCOL_PARAMETERS, PREPROD_GLOBAL_PARAMETERS, PREVIEW_DEFAULT_PROTOCOL_PARAMETERS,
-        PREVIEW_GLOBAL_PARAMETERS, ProtocolParameters,
+        MAINNET_DEFAULT_PROTOCOL_PARAMETERS, PREPROD_DEFAULT_PROTOCOL_PARAMETERS, PREVIEW_DEFAULT_PROTOCOL_PARAMETERS,
+        ProtocolParameters,
     },
     protocol_parameters_update::{ProtocolParamUpdate, display_protocol_parameters_update},
     protocol_version::{self, PROTOCOL_VERSION_10, ProtocolVersion},
